clientEngine impl. for recoverPassword method
