+++ progetto/ClientProject/src/client/Client.java
@@ -176,33 +176,7 @@ public class Client extends AnonymousClient {
         return false;
     }
 
-    /**
-     * Permette all'utente di creare una nuova password per l'account
-     * @param email l'email associata all'account
-     * @param newPassword
-     * @param repeatPassword
-     * @return
-     */
-    public boolean recoverPassword(String email,String newPassword,String repeatPassword){
-        try {
-            ResponseCode resp = server_stub.recoverPassword(email, newPassword, repeatPassword, this.skeleton);
-            if(resp.IsOK()){
-                infoStamp("password successfully changed.");
-                return true;
-            }else{
-                if(resp.getCodice()==ResponseCode.Codici.R510){
-                    infoStamp("invalid arguments.");
-                }else{
-                    infoStamp("Unknown error.");
-                }
-                return false;
-            }
-        }catch(Exception exc){
-            errorStamp("Not connected.");
-        }
 
-        return false;
-    }
 
     public String getPlainPassword() {
         return plainPassword;
